Refactoring more bits
