+++ datavec-spark/pom.xml
@@ -34,6 +34,7 @@
           <version>1.12</version>
           <executions>
             <execution>
+                <id>add-source</id>
               <phase>generate-sources</phase>
               <goals><goal>add-source</goal></goals>
               <configuration>
@@ -42,8 +43,18 @@
                 </sources>
               </configuration>
             </execution>
+            <execution>
+              <id>add-test-source</id>
+              <phase>generate-test-sources</phase>
+              <goals><goal>add-test-source</goal></goals>
+              <configuration>
+                  <sources>
+                      <source>src/test/spark-${spark.major.version}</source>
+                  </sources>
+              </configuration>
+            </execution>
           </executions>
-        </plugin>        
+        </plugin>
       </plugins>
     </build>
