+++ datavec-spark/src/main/spark-2/org/datavec/spark/transform/BaseFlatMapFunctionAdaptee.java
@@ -1,0 +1,19 @@
+package org.datavec.spark.transform;
+
+import java.util.Iterator;
+
+import org.apache.spark.api.java.function.FlatMapFunction;
+import org.datavec.spark.functions.FlatMapFunctionAdapter;
+
+public class BaseFlatMapFunctionAdaptee<K, V> implements FlatMapFunction<K,V> {
+
+    protected final FlatMapFunctionAdapter<K, V> adapter;
+
+    public BaseFlatMapFunctionAdaptee(FlatMapFunctionAdapter<K, V> adapter) {
+        this.adapter = adapter;
+    }
+    @Override
+    public Iterator<V> call(K k) throws Exception {
+        return adapter.call(k).iterator();
+    }
+}
