--- datavec-spark/src/main/spark-2/org/datavec/spark/transform/sparkfunction/SequenceToRows.java
@@ -1,46 +1,0 @@
-/*
- *  * Copyright 2016 Skymind, Inc.
- *  *
- *  *    Licensed under the Apache License, Version 2.0 (the "License");
- *  *    you may not use this file except in compliance with the License.
- *  *    You may obtain a copy of the License at
- *  *
- *  *        http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  *    Unless required by applicable law or agreed to in writing, software
- *  *    distributed under the License is distributed on an "AS IS" BASIS,
- *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  *    See the License for the specific language governing permissions and
- *  *    limitations under the License.
- */
-
-package org.datavec.spark.transform.sparkfunction;
-
-import org.apache.spark.api.java.function.FlatMapFunction;
-import org.apache.spark.api.java.function.Function;
-import org.apache.spark.sql.Row;
-import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema;
-import org.apache.spark.sql.types.StructType;
-import org.datavec.api.transform.schema.Schema;
-import org.datavec.api.writable.Writable;
-import org.datavec.spark.transform.DataFrames;
-
-import java.util.*;
-
-/**
- * Convert a record to a row
- * @author Adam Gibson
- */
-public class SequenceToRows implements FlatMapFunction<List<List<Writable>>,Row> {
-
-    private final SequenceToRowsAdapter adapter;
-
-    public SequenceToRows(Schema schema) {
-        this.adapter = new SequenceToRowsAdapter(schema);
-    }
-
-    @Override
-    public Iterator<Row> call(List<List<Writable>> sequence) throws Exception {
-        return adapter.call(sequence).iterator();
-    }
-}
