+++ datavec-spark/src/test/spark-2/org/datavec/spark/transform/NormalizationTests.java
@@ -4,7 +4,6 @@ import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.datavec.api.transform.schema.Schema;
-import org.datavec.api.transform.transform.normalize.Normalize;
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.Writable;
 import org.datavec.common.RecordConverter;
@@ -48,7 +47,7 @@ public class NormalizationTests extends BaseSparkTest {
 
         Schema schema = builder.build();
         JavaRDD<List<Writable>> rdd = sc.parallelize(data);
-        
+        Dataset<Row> dataFrame = DataFrames.toDataFrame(schema,rdd);
 
         //assert equivalent to the ndarray pre processing
         NormalizerStandardize standardScaler = new NormalizerStandardize();
@@ -59,7 +58,7 @@ public class NormalizationTests extends BaseSparkTest {
         zeroToOne.fit(new DataSet(arr.dup(),arr.dup()));
         INDArray zeroToOnes = arr.dup();
         zeroToOne.transform(new DataSet(zeroToOnes,zeroToOnes));
-        List<Row> rows = Normalization.stdDevMeanColumns(DataFrames.toDataFrame(schema,rdd), DataFrames.toDataFrame(schema,rdd).columns());
+        List<Row> rows = Normalization.stdDevMeanColumns(dataFrame,dataFrame.columns());
         INDArray assertion = DataFrames.toMatrix(rows);
         //compare standard deviation
         assertTrue(standardScaler.getStd().equalsWithEps(assertion.getRow(0),1e-1));
@@ -94,9 +93,10 @@ public class NormalizationTests extends BaseSparkTest {
         assertEquals(schema,DataFrames.fromStructType(DataFrames.fromSchema(schema)));
         assertEquals(rdd.collect(),DataFrames.toRecords(DataFrames.toDataFrame(schema,rdd)).getSecond().collect());
 
-        DataFrames.toDataFrame(schema,rdd).show();
-        Normalization.zeromeanUnitVariance(DataFrames.toDataFrame(schema,rdd)).show();
-        Normalization.normalize(DataFrames.toDataFrame(schema,rdd)).show();
+        Dataset<Row> dataFrame = DataFrames.toDataFrame(schema,rdd);
+        dataFrame.show();
+        Normalization.zeromeanUnitVariance(dataFrame).show();
+        Normalization.normalize(dataFrame).show();
 
         //assert equivalent to the ndarray pre processing
         NormalizerStandardize standardScaler = new NormalizerStandardize();
